pipeline/requirements.txt
supabase==2.15.0
requests==2.32.3
python-dotenv==1.0.1
python-dateutil==2.9.0
